Rename Sequence Player files
