+++ axelrod/strategies/sequence_player.py
