+++ axelrod/tests/unit/test_sequence_player.py
